backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
qdrant-client==1.11.0
sentence-transformers==3.1.1
google-genai==1.62.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1